test: :white_check_mark: Add test for checking `Return`
There is already a test for `hooksff.Return`, but that only runs the function `run_hooks_for`, so `src/hooksff/__init__.py:268` would not be ran in any tests, thus resulting in a lower coverage. This test does run the original function, which will run that line, thus resulting in a higher coverage.